refactor: prepare req/res body modules for extensibility
