+++ src/body/response/mod.rs
@@ -1,0 +1,6 @@
+mod body;
+mod stream_adapter;
+
+use stream_adapter::StreamAdapter;
+
+pub use body::Body;
