+++ src/request/mod.rs
@@ -1,4 +1,3 @@
-mod body;
 mod path_param;
 mod query_param;
 mod query_parser;
@@ -7,7 +6,6 @@ mod request;
 use query_parser::parse_query_params;
 
 pub use self::{
-    body::{Body, BodyStream},
     path_param::PathParam,
     query_param::{QueryParamValue, QueryParamValues, QueryParamValuesIter},
     request::Request,
