+++ src/response/mod.rs
@@ -1,10 +1,9 @@
-mod body;
 mod builder;
 mod into_response;
 mod redirect;
 mod response;
 
-pub use self::{
-    body::Body, builder::ResponseBuilder, into_response::IntoResponse, redirect::Redirect,
-    response::Response,
-};
+pub use builder::ResponseBuilder;
+pub use into_response::IntoResponse;
+pub use redirect::Redirect;
+pub use response::Response;
